requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
dj-database-url==0.5.0
# Django==4.0.4
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==5.1.0
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
gunicorn==20.1.0
Pillow==9.1.0
promise==2.3
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==6.0.0
